+++ src/main/java/com/alkemy/ong/web/controllers/RestExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.web.controller;
+package com.alkemy.ong.web.controllers;
 
 
 import com.alkemy.ong.web.exception.ParamNotFound;
